+++ src/jespermattiasson/Main.java
@@ -1,0 +1,13 @@
+package jespermattiasson;
+
+public class Main {
+
+	public static void main(String[] args) {
+		
+		// Create object for GUI.
+		UserInterface frame = new UserInterface();
+		// Display GUI.
+		frame.launchFrame();
+	}
+
+}
